+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.12
+
+* fix decision cards creation
+* remove multi account UI
+* 
+
 ## 0.0.11
 
 * fix missing updates